+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Renamed constraint method `required_constraints` to `parameterized_constraints`
 - Renamed Bifrost method `mocked` to `validation_only`
 - All exceptions include a `ctx` property for debugging
+- Mysql `INTERVAL` syntax support
 
 ## 0.3.0 - 7/15/23